Start timer on toggle on
Because of a bug in the ValueStepper library, the stepper value
always equals 0; therefore, the timer will always default to
the maximum time allowed.

The intended behavior is for the timer to run for the selected
number of minutes or if a time is not selected (i.e. stepper is 0),
then the the zone will run for the maximum time.